refactor(di): added koin extension to safely close resources - added missed lettuce dependency - moved resource closing logic from Application to DI - made all services/clients implement AutoClosable and configures onClose in KoinExtensions
