+++ ktor-service/src/main/kotlin/Application.kt
@@ -1,10 +1,9 @@
 package com.pavelryzh
 
-import com.pavelryzh.kafka.KafkaTrafficProducer
 import com.pavelryzh.plugins.*
+import com.pavelryzh.plugins.di.configureDI
 import io.ktor.server.application.*
 import io.ktor.server.netty.*
-import org.koin.ktor.ext.get
 
 fun main(args: Array<String>) {
     EngineMain.main(args)
@@ -16,10 +15,4 @@ fun Application.module() {
     configureKafka()
     configureDI()
     configureRouting()
-
-    monitor.subscribe(ApplicationStopped) {
-        val kafkaProducer = get<KafkaTrafficProducer>()
-        kafkaProducer.close()
-        log.info("KafkaTrafficProducer closed successfully.")
-    }
 }
