+++ spring-backend/src/main/resources/application.properties
@@ -8,3 +8,4 @@ spring.flyway.locations=classpath:db/migration
 logging.level.org.flywaydb=DEBUG
 logging.level.org.springframework.boot.autoconfigure.flyway=DEBUG
 spring.flyway.fail-on-missing-locations=true
+spring.threads.virtual.enabled=true
